+++ readme.md
@@ -5,11 +5,11 @@ Bot che gestisce la tua lista di film e che sceglie per te cosa vedere quando se
 ### Versione 0.2 (File-inator)
 Cominciamo a salvare roba
 #### Novità
-- [ ] _Comando /file_: il bot ti presenta una serie di opzioni:
-	- [ ] _Nuovo File_: crea un nuovo file in cui salvare robe (per ora solo frasi);
-	- [ ] _Salva Roba: salva in un file;
-	- [ ] _Scancellamento: cancella file.
-- [ ] _Comando /cancel_: Annulla l'ultima operazione
+- _Comando /file_: il bot ti presenta una serie di opzioni:
+	- _Nuovo File_: crea un nuovo file in cui salvare robe (per ora solo frasi);
+	- _Salva Roba_: salva in un file;
+	- _Scancellamento_: cancella file.
+- _Comando /cancel_: Annulla l'ultima operazione
 #### Comandi
 - */start*: Manda un saluto all'utente;
 - */file*: Tutto ciò che serve per salvare le robe utili;
